+++ src/services/jokeService.ts
@@ -2,27 +2,9 @@ import type { Joke } from "@/interfaces/joke.interface";
 
 const BASE_URL = import.meta.env.VITE_JOKE_API_BASE_URL;
 
-export async function getRandomJoke(): Promise<Joke> {
-  const res = await fetch(`${BASE_URL}/random_joke`);
-  if (!res.ok) throw new Error("Error fetching joke");
-  return await res.json();
-}
-
-export async function getRandomJokes(count: number): Promise<Joke[]> {
-  const res = await fetch(`${BASE_URL}/jokes/random/${count}`);
-  if (!res.ok) throw new Error("Error fetching jokes");
-  return await res.json();
-}
-
 export async function getJokeTypes(): Promise<string[]> {
   const res = await fetch(`${BASE_URL}/types`);
   if (!res.ok) throw new Error("Error fetching joke types");
-  return await res.json();
-}
-
-export async function getJokeById(id: number): Promise<Joke> {
-  const res = await fetch(`${BASE_URL}/jokes/${id}`);
-  if (!res.ok) throw new Error("Error fetching joke");
   return await res.json();
 }