+++ src/utils/index.ts
@@ -1,10 +1,14 @@
 import { CACHE_DURATION, JOKES_CACHE_KEY, TYPES_CACHE_KEY } from "@/const";
-import type { JokesCacheData, TypesCacheData } from "@/interfaces/cache-data.interface";
+import type {
+  JokesCacheData,
+  TypesCacheData,
+} from "@/interfaces/cache-data.interface";
 
 export const confirmModal = (title: string) => window.confirm(title);
 
-
-export function saveToLocalStorage<T extends TypesCacheData["data"] | JokesCacheData["data"]>(key: string, data: T) {
+export function saveToLocalStorage<
+  T extends TypesCacheData["data"] | JokesCacheData["data"],
+>(key: string, data: T) {
   try {
     const cacheData = {
       data,
@@ -16,7 +20,9 @@ export function saveToLocalStorage<T extends TypesCacheData["data"] | JokesCache
   }
 }
 
-export function loadJokesFromLocalStorage(key: string): JokesCacheData["data"] | null {
+export function loadJokesFromLocalStorage(
+  key: string
+): JokesCacheData["data"] | null {
   try {
     const cached = localStorage.getItem(key);
     if (!cached) return null;
@@ -35,7 +41,9 @@ export function loadJokesFromLocalStorage(key: string): JokesCacheData["data"] |
     return null;
   }
 }
-export function loadTypesFromLocalStorage(key: string): TypesCacheData["data"] | null {
+export function loadTypesFromLocalStorage(
+  key: string
+): TypesCacheData["data"] | null {
   try {
     const cached = localStorage.getItem(key);
     if (!cached) return null;
@@ -56,6 +64,6 @@ export function loadTypesFromLocalStorage(key: string): TypesCacheData["data"] |
 }
 
 export function clearCache() {
-    localStorage.removeItem(JOKES_CACHE_KEY);
-    localStorage.removeItem(TYPES_CACHE_KEY);
-  }
+  localStorage.removeItem(JOKES_CACHE_KEY);
+  localStorage.removeItem(TYPES_CACHE_KEY);
+}